move pkpanda and cosmos integration tests to test_packaging.py
