+++ packages/dcos-integration-test/extra/test_applications.py
@@ -237,62 +237,3 @@ def octarine_runner(dcos_api_session, mode, uuid, uri, bind_port=None):
 
     with dcos_api_session.marathon.deploy_and_cleanup(app):
         pass
-
-
-def test_pkgpanda_api(dcos_api_session):
-
-    def get_and_validate_package_ids(path, node):
-        r = dcos_api_session.get(path, node=node)
-        assert r.status_code == 200
-        package_ids = r.json()
-        assert isinstance(package_ids, list)
-        for package_id in package_ids:
-            r = dcos_api_session.get(path + package_id, node=node)
-            assert r.status_code == 200
-            name, version = package_id.split('--')
-            assert r.json() == {'id': package_id, 'name': name, 'version': version}
-        return package_ids
-
-    active_buildinfo = dcos_api_session.get('/pkgpanda/active.buildinfo.full.json').json()
-    active_buildinfo_packages = sorted(
-        # Setup packages don't have a buildinfo.
-        (package_name, info['package_version'] if info else None)
-        for package_name, info in active_buildinfo.items()
-    )
-
-    def assert_packages_match_active_buildinfo(package_ids):
-        packages = sorted(map(lambda id_: tuple(id_.split('--')), package_ids))
-        assert len(packages) == len(active_buildinfo_packages)
-        for package, buildinfo_package in zip(packages, active_buildinfo_packages):
-            if buildinfo_package[1] is None:
-                # No buildinfo for this package, so we can only compare names.
-                assert package[0] == buildinfo_package[0]
-            else:
-                assert package == buildinfo_package
-
-    for node in dcos_api_session.masters + dcos_api_session.all_slaves:
-        package_ids = get_and_validate_package_ids('pkgpanda/repository/', node)
-        active_package_ids = get_and_validate_package_ids('pkgpanda/active/', node)
-
-        assert set(active_package_ids) <= set(package_ids)
-        assert_packages_match_active_buildinfo(active_package_ids)
-
-
-def test_packaging_api(dcos_api_session):
-    """Test the Cosmos API (/package) wrapper
-    """
-    install_response = dcos_api_session.cosmos.install_package('kafka', '1.1.9-0.10.0.0')
-    data = install_response.json()
-
-    dcos_api_session.marathon.poll_marathon_for_app_deployment(data['appId'], 1,
-                                                               True, False)
-
-    list_response = dcos_api_session.cosmos.list_packages()
-    packages = list_response.json()['packages']
-    assert len(packages) == 1 and packages[0]['appId'] == data['appId']
-
-    dcos_api_session.cosmos.uninstall_package('kafka', data['appId'])
-
-    list_response = dcos_api_session.cosmos.list_packages()
-    packages = list_response.json()['packages']
-    assert len(packages) == 0
